+++ sources/CHANGELOG.md
@@ -5,6 +5,7 @@
  * Improves panel alignment when in overflow menu (#73).
  * Adds date to exported file names (#74).
  * Adds shortcut for privately bookmarking all tabs in the current window at once (#80).
+ * Allows custom folder titles (#85).
 
 ### 0.1.16
  * Adds the ability to customize command shortcuts starting from Firefox 60 (#55).